requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
distlib==0.3.5
dj-database-url==1.0.0
Django==4.2.1
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
gunicorn==20.1.0
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.2.1
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.10
tzdata==2022.1
virtualenv==20.16.3
whitenoise==6.2.0
